requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
streamlit==1.27.2
Pillow==9.5.0
